requirements.txt
numpy==1.15.2
pandas==0.23.4
scikit-learn==0.20.0
scipy==1.1.0
sklearn==0.0
flask==1.0.2
python-dotenv==0.9.1
